Login Page Implemented
Main Activity is now a login page. When the user enters the correct
credentials, the Quest Board Activity starts (it's currently empty).
Otherwise the user is warned about incorrect input. Currently it checks
if the username field is empty, then the password field, else it
determines if the two together match the required credentials.