Parse Integration for Login
Modified the checking on the login fields to properly display errors for
both fields if either is empty.
Created the App class to take care of initializing the database
connection for every activity.
Registration Activity also has been implemented, but it doesn't check if
the username already exists in the database.