Settings Implementation and Password Encryption
Removed the sending of the password field from the Main to the
Registration Activity.
Registration now properly ensures the username is not already in the
database.
The Main Activity's label has been set to "app_name" because the label
for the application seems to be ignored in favor of Main's label.
Settings Activity now implemented and populates/saves the Name, Location
of Origin, and Alignment fields/spinner.
Passwords are now encrypted.
Properly breaks out of the loop when attempting to log in and the
username has been found but the password is incorrect.